rename showroom to CinemaHall
